Restructure to make features folder more intuitive and decreased build size from 17.5 mb to 11
