+++ features/degree-audit-app/shared/audit-card.tsx
